nebulas/requirements.txt
bcrypt==4.0.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
PyYAML==6.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
